+++ projects/ngx-highlightjs/src/lib/highlight.service.ts
@@ -1,26 +1,33 @@
 import { Injectable, Inject, Optional } from '@angular/core';
-import { HighlightConfig, HighlightResult, HighlightLanguage, HighlightOptions, HIGHLIGHT_OPTIONS } from './highlight.model';
-import hljs from 'highlight.js/lib/highlight.js';
+import { Observable } from 'rxjs';
+import { filter, map, tap } from 'rxjs/operators';
+import { HighlightConfig, HighlightResult, HighlightLibrary, HighlightOptions, HIGHLIGHT_OPTIONS } from './highlight.model';
+import { HighlightLoader } from './highlight.loader';
 
 @Injectable({
   providedIn: 'root'
 })
 export class HighlightJS {
-  constructor(@Optional() @Inject(HIGHLIGHT_OPTIONS) options: HighlightOptions) {
-    if (options) {
-      // Register HighlightJS languages
-      options.languages().map((language: HighlightLanguage) =>
-        this.registerLanguage(language.name, language.func)
-      );
-      if (options.config) {
+
+  private _hljs!: HighlightLibrary;
+
+  // A reference for hljs library
+  get hljs(): HighlightLibrary | null {
+    return this._hljs;
+  }
+
+  constructor(private _loader: HighlightLoader, @Optional() @Inject(HIGHLIGHT_OPTIONS) options: HighlightOptions) {
+    // Load highlight.js library on init
+    _loader.ready.pipe().subscribe((hljs: HighlightLibrary) => {
+      this._hljs = hljs;
+      if (options && options.config) {
         // Set global config if present
-        this.configure(options.config);
+        hljs.configure(options.config);
+        if (hljs.listLanguages().length < 1) {
+          console.error('[HighlightJS]: No languages were registered!');
+        }
       }
-    }
-    // Throw an error if no languages were registered.
-    if (this.listLanguages().length < 1) {
-      throw new Error('[HighlightJS]: No languages were registered!');
-    }
+    });
   }
 
   /**
@@ -32,8 +39,10 @@ export class HighlightJS {
    * @param continuation An optional mode stack representing unfinished parsing.
    * When present, the function will restart parsing from this state instead of initializing a new one
    */
-  highlight(name: string, value: string, ignore_illegals: boolean, continuation?: any): HighlightResult {
-    return hljs.highlight(name, value, ignore_illegals, continuation);
+  highlight(name: string, value: string, ignore_illegals: boolean, continuation?: any): Observable<HighlightResult> {
+    return this._loader.ready.pipe(
+      map((hljs: HighlightLibrary) => hljs.highlight(name, value, ignore_illegals, continuation))
+    );
   }
 
   /**
@@ -42,8 +51,10 @@ export class HighlightJS {
    * @param languageSubset An optional array of language names and aliases restricting detection to only those languages.
    * The subset can also be set with configure, but the local parameter overrides the option if set.
    */
-  highlightAuto(value: string, languageSubset: string[]): HighlightResult {
-    return hljs.highlightAuto(value, languageSubset);
+  highlightAuto(value: string, languageSubset: string[]): Observable<HighlightResult> {
+    return this._loader.ready.pipe(
+      map((hljs: HighlightLibrary) => hljs.highlightAuto(value, languageSubset))
+    );
   }
 
   /**
@@ -52,8 +63,10 @@ export class HighlightJS {
    * Options are set globally with configure.
    * @param value Accepts a string with the highlighted markup
    */
-  fixMarkup(value: string): string {
-    return hljs.fixMarkup(value);
+  fixMarkup(value: string): Observable<string> {
+    return this._loader.ready.pipe(
+      map((hljs: HighlightLibrary) => hljs.fixMarkup(value))
+    );
   }
 
   /**
@@ -62,23 +75,29 @@ export class HighlightJS {
    * See the class reference for all available language names and aliases.
    * @param block The element to apply highlight on.
    */
-  highlightBlock(block: HTMLElement) {
-    hljs.highlightBlock(block);
+  highlightBlock(block: HTMLElement): Observable<void> {
+    return this._loader.ready.pipe(
+      map((hljs: HighlightLibrary) => hljs.highlightBlock(block))
+    );
   }
 
   /**
    * Configures global options:
-   * @param config
+   * @param config HighlightJs configuration argument
    */
-  configure(config: HighlightConfig) {
-    hljs.configure(config);
+  configure(config: HighlightConfig): Observable<void> {
+    return this._loader.ready.pipe(
+      map((hljs: HighlightLibrary) => hljs.configure(config))
+    );
   }
 
   /**
    * Applies highlighting to all <pre><code>..</code></pre> blocks on a page.
    */
-  initHighlighting() {
-    hljs.initHighlighting();
+  initHighlighting(): Observable<void> {
+    return this._loader.ready.pipe(
+      map((hljs: HighlightLibrary) => hljs.initHighlighting())
+    );
   }
 
   /**
@@ -87,15 +106,19 @@ export class HighlightJS {
    * @param language A function that returns an object which represents the language definition.
    * The function is passed the hljs object to be able to use common regular expressions defined within it.
    */
-  registerLanguage(name: string, language: Function) {
-    hljs.registerLanguage(name, language);
+  registerLanguage(name: string, language: () => any): Observable<HighlightLibrary> {
+    return this._loader.ready.pipe(
+      tap((hljs: HighlightLibrary) => hljs.registerLanguage(name, language))
+    );
   }
 
   /**
    * @return The languages names list.
    */
-  listLanguages(): string[] {
-    return hljs.listLanguages();
+  listLanguages(): Observable<string[]> {
+    return this._loader.ready.pipe(
+      map((hljs: HighlightLibrary) => hljs.listLanguages())
+    );
   }
 
   /**
@@ -103,7 +126,20 @@ export class HighlightJS {
    * @param name Language name
    * @return The language object if found, undefined otherwise.
    */
-  getLanguage(name: string): any {
-    return hljs.getLanguage(name);
+  getLanguage(name: string): Observable<any> {
+    return this._loader.ready.pipe(
+      map((hljs: HighlightLibrary) => hljs.getLanguage(name))
+    );
+  }
+
+  /**
+   * Display line numbers
+   * @param el Code element
+   */
+  lineNumbersBlock(el: HTMLElement): Observable<any> {
+    return this._loader.ready.pipe(
+      filter((hljs: HighlightLibrary) => !!hljs.lineNumbersBlock),
+      tap((hljs: HighlightLibrary) => hljs.lineNumbersBlock(el))
+    );
   }
 }