+++ src/app/material/material.module.ts
@@ -1,34 +1,27 @@
 import { NgModule } from '@angular/core';
-import {
-  MatButtonModule,
-  MatGridListModule,
-  MatIconModule,
-  MatInputModule,
-  MatListModule,
-  MatButtonToggleModule,
-  MatCardModule,
-  MatChipsModule,
-  MatAutocompleteModule,
-  MatToolbarModule,
-  MatSelectModule
-} from '@angular/material';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { MatCardModule } from '@angular/material/card';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatSelectModule } from '@angular/material/select';
+import { MatTabsModule } from '@angular/material/tabs';
 
 const MATERIAL_MODULES = [
-  MatGridListModule,
-  MatInputModule,
   MatButtonModule,
   MatListModule,
   MatIconModule,
-  MatButtonToggleModule,
   MatCardModule,
-  MatChipsModule,
-  MatAutocompleteModule,
   MatToolbarModule,
-  MatSelectModule
+  MatSelectModule,
+  MatTabsModule,
+  MatProgressBarModule
 ];
 
 @NgModule({
   imports: [MATERIAL_MODULES],
   exports: [MATERIAL_MODULES]
 })
-export class MaterialModule { }
+export class MaterialModule {
+}